+++ packages/components-react/src/examples/CarouselExampleFocusOnCenterSlide.tsx
@@ -40,6 +40,10 @@ export const CarouselExampleFocusOnCenterSlidePage = (): JSX.Element => {
       heading="Some Heading"
       onUpdate={(e) => onCarouselUpdate(e.detail.activeIndex)}
       focusOnCenterSlide
+      trimSpace={false}
+      gradientColor="background-surface"
+      // @ts-expect-error: ignore unknown property (TS2353)
+      style={{ '--p-gradient-color-width': '25%' }}
     >
       {Array.from({ length: 6 }).map((_, i) => (
         <div style={getSlideStyle(i)} key={i}>