+++ packages/components-vue/src/examples/CarouselExampleFocusOnCenterSlide.vue
@@ -20,8 +20,10 @@
 
 <template>
   <PCarousel
+    gradientColor="background-surface"
     :slidesPerPage="3"
     :focusOnCenterSlide="true"
+    :trimSpace="false"
     :intl="{
       slideLabel: 'Slide %s von %s',
       prev: 'Vorheriger Slide',
@@ -31,6 +33,7 @@
     }"
     :heading="'Some heading'"
     @update="onCarouselUpdate"
+    :style="{ '--p-gradient-color-width': '25%' }"
   >
     <div v-for="(_, i) in Array.from(Array(6))" :key="i" :class="getSlideClass(i)">Slide {{ i + 1 }}</div>
   </PCarousel>